+++ examples/data-sources/sdwan_device_acl_policy_definition/data-source.tf
@@ -1,3 +1,0 @@
-data "sdwan_device_acl_policy_definition" "example" {
-  id = "f6b2c44c-693c-4763-b010-895aa3d236bd"
-}