+++ examples/resources/sdwan_device_acl_policy_definition/import.sh
@@ -1,1 +1,0 @@
-terraform import sdwan_device_acl_policy_definition.example "f6b2c44c-693c-4763-b010-895aa3d236bd"