+++ examples/resources/sdwan_ipv4_device_acl_policy_definition/resource.tf
@@ -1,11 +1,10 @@
-resource "sdwan_device_acl_policy_definition" "example" {
+resource "sdwan_ipv4_device_acl_policy_definition" "example" {
   name           = "Example"
   description    = "My description"
   default_action = "drop"
   sequences = [
     {
       id          = 10
-      ip_type     = "ipv4"
       name        = "Sequence 10"
       base_action = "accept"
       match_entries = [